refactor: 404 응답 처리
