+++ tomcat/src/main/java/com/techcourse/exception/NotFoundException.java
@@ -1,0 +1,8 @@
+package com.techcourse.exception;
+
+public class NotFoundException extends RuntimeException {
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
