+++ tomcat/src/main/java/org/apache/coyote/http11/response/StatusCode.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
 
 public enum StatusCode {
     OK(200, "OK"),
